refactor: change sequence of fields in the react form
